Moved Shadow DOM tests. Added focus and selection support for shadow DOM in setup(). Removed old custom elements test
