+++ tests/clipboard/paste.ts
@@ -1,6 +1,8 @@
+import type {ShadowInput} from '../_helpers/shadow-input'
 import userEvent from '#src'
 import {render, setup} from '#testHelpers'
 import {createDataTransfer} from '#src/utils'
+import '../_helpers/shadow-input'
 
 test('paste with empty clipboard', async () => {
   const {element, getEvents, user} = setup(`<input/>`)
@@ -147,5 +149,15 @@ describe('without Clipboard API', () => {
       `[Error: \`userEvent.paste()\` without \`clipboardData\` requires the \`ClipboardAPI\` to be available.]`,
     )
     expect(getEvents()).toHaveLength(0)
+  })
+})
+
+describe('on shadow DOM', () => {
+  test('paste into an input element', async () => {
+    const {element, user} = setup('<shadow-input></shadow-input>')
+
+    await user.paste('test')
+
+    expect((element as ShadowInput).value).toEqual('test')
   })
 })
